% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENB_PERFECT.R
\name{enb_perfect}
\alias{enb_perfect}
\title{enb_perfect}
\usage{
enb_perfect(D, U, Theta, t, prop, cost)
}
\arguments{
\item{D}{Number of decision options}

\item{U}{Utility function that depends on the decision option, parameters and decision times t_1 and t_2}

\item{Theta}{Named matrix of parameter draws from prior/posterior distribution}

\item{t}{Named vector of ascending decision times in years including the current time ("C"), analysis time ("A") and the time horizon ("H")}

\item{prop}{Vector containing current proportions of intervention use. Must sum to one.}

\item{cost}{Cost of sampling}
}
\value{
Expected net benefit of collecting perfect information
}
\description{
Computes the expected net benefit of collecting perfect information using a Monte-Carlo approximation method
}
\examples{
# one parameter, two decision options
D <- 2
U <- function(d, Theta, t_1, t_2) sum(1.05^(1-(t_1:t_2)))*(-1)^(d-1)*(Theta - 0.4)
N <- 10000
Theta <- matrix(rbeta(N, 2, 3), nrow = N, ncol = 1, dimnames = list(NULL, "theta"))
t <- c(C = 0, A = 1, H = 15)
prop <- rep(1/D, D)
cost <- 0
enb_perfect(D, U, Theta, t, prop, cost)

# two parameters, two decision options
D <- 2
U <- function(d, Theta, t_1, t_2) sum(1.05^(1-(t_1:t_2)))*Theta[,d]
N <- 10000
Theta <- matrix(c(rbeta(N, 2, 3), rbeta(N, 2, 3)),
                nrow = N, ncol = 2, dimnames = list(NULL, c("theta_A", "theta_B")))
t <- c(C = 0, A = 1, H = 15)
prop <- rep(1/D, D)
cost <- 0
enb_perfect(D, U, Theta, t, prop, cost)

# three parameters, three decision options
D <- 3
U <- function(d, Theta, t_1, t_2) sum(1.05^(1-(t_1:t_2)))*Theta[,d]
N <- 10000
Theta <- matrix(c(rbeta(N, 2, 3), rbeta(N, 2, 3), rbeta(N, 2, 3)),
                nrow = N, ncol = 3, dimnames = list(NULL, c("theta_A", "theta_B", "theta_C")))
t <- c(C = 0, A = 1, H = 15)
prop <- rep(1/D, D)
cost <- 0
enb_perfect(D, U, Theta, t, prop, cost)
}
