% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EVPI.R
\name{evpi}
\alias{evpi}
\title{evpi}
\usage{
evpi(D, U, Theta)
}
\arguments{
\item{D}{Number of decision options}

\item{U}{Utility function that depends on the decision option and parameters}

\item{Theta}{Named matrix of parameter draws from prior/posterior distribution}
}
\value{
Expected value of perfect information
}
\description{
Computes the expected value of perfect information use a Monte-Carlo approximation
}
\examples{
# one parameter, two decision options
D <- 2
U <- function(d, Theta) (-1)^(d-1)*(Theta - 0.4)
N <- 10000
Theta <- matrix(rbeta(N, 2, 3), nrow = N, ncol = 1, dimnames = list(NULL, "theta"))
evpi(D, U, Theta)

# two parameters, two decision options
D <- 2
U <- function(d, Theta) Theta[,d]
N <- 10000
Theta <- matrix(c(rbeta(N, 2, 3), rbeta(N, 2, 3)),
                nrow = N, ncol = 2, dimnames = list(NULL, c("theta_A", "theta_B")))
evpi(D, U, Theta)

# three parameters, three decision options
D <- 3
U <- function(d, Theta) Theta[,d]
N <- 10000
Theta <- matrix(c(rbeta(N, 2, 3), rbeta(N, 2, 3), rbeta(N, 2, 3)),
                nrow = N, ncol = 3, dimnames = list(NULL, c("theta_A", "theta_B", "theta_C")))
evpi(D, U, Theta)
}
