% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_betabinomial_trial.R
\name{sim_betabinomial_trial}
\alias{sim_betabinomial_trial}
\title{sim_betabinomial_trial}
\usage{
sim_betabinomial_trial(
  D,
  U,
  Theta,
  prior,
  n_analyses = 1,
  n_samp = 1,
  n_sims = 1,
  n_draws = 10000,
  method = "NP",
  J = 10000,
  K = 10000,
  n_cores = 1
)
}
\arguments{
\item{D}{Number of decision options}

\item{U}{Utility function that depends on the decision option and parameters}

\item{Theta}{Vector of true parameter values to use for simulation. Must be of length D.}

\item{prior}{Matrix of parameters to specify the beta prior distribution with each row representing a different parameter contained in Theta.}

\item{n_analyses}{Maximum number of trial analyses (must be at least two, including the final analysis)}

\item{n_samp}{Number of observations to sample between analyses for each decision option}

\item{n_sims}{Number of simulated trials to generate (must be at least two)}

\item{n_draws}{Number of parameter samples to draw from prior/posterior distributions}

\item{method}{Approximation method. Either MC for Monte-Carlo or NP for non-parametric (default). The moment matching method is not currently supported.}

\item{J}{Number of inner Monte Carlo loops. Only required for the Monte Carlo approximation method.}

\item{K}{Number of outer Monte Carlo loops. Only required for the Monte Carlo approximation method.}

\item{n_cores}{Number of cores to use if running in parallel}
}
\value{
Matrix containing the expected value of sample information computed at each analysis for each simulated trial
}
\description{
Simulates a series of trials using the value-driven adaptive design for a beta-binomial model
}
\examples{
D <- 2
U <- function(d, Theta) Theta[,d]
Theta <- c(0.5, 0.6)
prior <- matrix(rep(c(2, 3), D), nrow = D, byrow = TRUE)
sim_betabinomial_trial(D = D, U = U, Theta = Theta, prior = prior,
                       n_analyses = 3, n_samp = 50, n_sims = 4)
}
