% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValueAdapt-package.R
\docType{package}
\name{ValueAdapt-package}
\alias{ValueAdapt}
\alias{ValueAdapt-package}
\title{ValueAdapt: The Value-Driven Adaptive Design}
\description{
Contains functions to compute value of information metrics applied to an adaptive trial design.
}
\author{
\strong{Maintainer}: Michael Dymock \email{michael.dymock1@hotmail.com} (\href{https://orcid.org/0000-0003-4881-2860}{ORCID})

}
\keyword{internal}
