% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIM_BETABINOMIAL_TRIAL.R
\name{sim_betabinomial_trial}
\alias{sim_betabinomial_trial}
\title{sim_betabinomial_trial}
\usage{
sim_betabinomial_trial(
  D,
  U,
  Theta,
  t,
  prop,
  cost,
  prior_par,
  n_analyses = 1,
  n_samp = 1,
  n_sims = 1,
  n_draws = 10000,
  method = "NP",
  J = 10000,
  K = 10000,
  n_cores = 1
)
}
\arguments{
\item{D}{Number of decision options}

\item{U}{Utility function that depends on the decision option and parameters}

\item{Theta}{Vector of true parameter values to use for simulation. Must be of length D.}

\item{t}{Named matrix of ascending decision times in years including the current time ("C"), analysis time ("A") and the time horizon ("H") for each analysis including the trial start-up decision. Columns are times and rows are analyses.}

\item{prop}{Vector containing current proportions of intervention use. Must sum to one.}

\item{cost}{Vector of costs to proceed between analyses. Must be of length n_analyses + 1 to include the trial start-up decision.}

\item{prior_par}{Matrix of parameters to specify the beta prior distribution with each row representing a different parameter contained in Theta.}

\item{n_analyses}{Maximum number of trial analyses where ENBS decisions are made (must be at least one)}

\item{n_samp}{Number of observations to sample between analyses for each decision option}

\item{n_sims}{Number of simulated trials to generate (must be at least two)}

\item{n_draws}{Number of parameter samples to draw from prior/posterior distributions}

\item{method}{Approximation method. Either MC for Monte-Carlo or NP for non-parametric (default). The moment matching method is not currently supported.}

\item{J}{Number of inner Monte Carlo loops. Only required for the Monte Carlo approximation method.}

\item{K}{Number of outer Monte Carlo loops. Only required for the Monte Carlo approximation method.}

\item{n_cores}{Number of cores to use if running in parallel}
}
\value{
Matrix containing the expected value of sample information computed at each analysis for each simulated trial
}
\description{
Simulates a series of trials using the value-driven adaptive design for a beta-binomial model
}
\examples{
D <- 2
U <- function(d, Theta, t_1, t_2) sum(1.05^(1-(t_1:t_2)))*Theta[,d]
Theta <- c(0.5, 0.6)
n_analyses <- 2
t <- matrix(c(0, 1, 15,
              1, 2, 15,
              2, 3, 15),
            nrow = n_analyses + 1, ncol = 3, byrow = TRUE,
            dimnames = list(NULL, c("C", "A", "H")))
prop <- rep(1/D, D)
cost <- rep(0, n_analyses + 1)
prior_par <- matrix(rep(c(2, 3), D), nrow = D, byrow = TRUE)
sim_betabinomial_trial(D, U, Theta, t, prop, cost, prior_par,
                       n_analyses = n_analyses, n_samp = 50,
                       n_sims = 4)
}
